Added some sprites and rebinded input actions

+++ Assets/Scripts/Core/ECS/Components/Events/Units/HitEventComponent.cs
@@ -4,7 +4,7 @@ namespace Core.ECS.Events
 {
     internal struct HitEventComponent
     {
-        public float Damage;
+        public byte Damage;
         public Vector2 HitPosition;
         public float HitRadius;
         public int TargetLayer;
